backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dateutil==2.9.0
google-api-python-client==2.155.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
apscheduler==3.10.4
alembic==1.14.0
icalendar==6.1.0
pyyaml==6.0.2
